+++ tests/bunit.web.tests/TestDoubles/Components/StubTest.cs
@@ -1,3 +1,6 @@
+using Bunit.TestAssets.SampleComponents.DisposeComponents;
+using Microsoft.AspNetCore.Components.Forms;
+
 #if NET5_0_OR_GREATER
 namespace Bunit.TestDoubles.Components;
 
@@ -25,6 +28,18 @@ public class StubTest : TestContext
 				ps => ps.ShouldContain(x => x.Key == nameof(Simple1.Header) && header.Equals(x.Value)),
 				ps => ps.ShouldContain(x => x.Key == nameof(Simple1.AttrValue) && attrValue.Equals(x.Value)),
 				ps => ps.Count.ShouldBe(2));
+	}
+
+	[Fact(DisplayName = "Stub<TComponent> can invoke event callbacks")]
+	public async Task Test003()
+	{
+		ComponentFactories.AddStub<InputText>();
+		var cut = RenderComponent<DisplayInput>();
+		var stubbedInput = cut.FindComponent<Stub<InputText>>();
+		
+		await stubbedInput.InvokeEventCallback(t => t.ValueChanged, "Hello World");
+		
+		cut.Find("p").MarkupMatches("<p>User wrote: Hello World</p>");
 	}
 }
 #endif